pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "justreadit"
version = "0.1.0"
description = "Hover-translation tool for Light.VN-based RPG games"
license = { text = "MPL-2.0" }
requires-python = ">=3.11"
dependencies = [
    # Windows OCR – split packages from winrt (Microsoft-maintained)
    "winrt-runtime",
    "winrt-Windows.Media.Ocr",
    "winrt-Windows.Graphics.Imaging",
    "winrt-Windows.Storage.Streams",
    "winrt-Windows.Foundation",
    "winrt-Windows.Foundation.Collections",
    "winrt-Windows.Globalization",
    "imagehash",
    "rapidfuzz",
    "pywin32",
    "psutil",
    "Pillow",
    "dxcam",        # DXGI Desktop Duplication capture
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pip-licenses",
    "pywin32-stubs",
]
ui = [
    "PySide6",
]
translators-cloud = [
    "google-cloud-translate>=3.0",
]
translators-free = [
    "deep-translator>=1.11",
]
translators-openai = [
    "openai>=1.0",
]
knowledge = [
    "mcp>=1.0",
    "numpy>=1.24",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ""
